+++ build.ts
@@ -1,0 +1,7 @@
+import { ChemicalBuild } from "chemicaljs";
+
+const build = new ChemicalBuild({
+	rammerhead: false,
+});
+
+await build.write();