+++ src/pages/apps.tsx
@@ -3,21 +3,21 @@ import { useTranslation } from "react-i18next";
 import { useState, useRef } from "preact/hooks";
 import { Web, setWeb } from "../components/web";
 import { useGlobalState } from "@ekwoka/preact-global-state";
-import { encodeURL } from "../util/searchURL";
 import { SearchIcon } from "../assets/searchIcon";
 import { CloseIcon } from "../assets/closeIcon";
 import apps from "../assets/apps.json";
 
 function Apps() {
     const { t } = useTranslation();
-    const [service] = useGlobalState<string>("service", localStorage.getItem("metallic/service") || "ultraviolet");
+    const [service] = useGlobalState<string>("service", localStorage.getItem("metallic/service") || "uv");
     const [webOpen, setWebOpen] = useState(false);
     const search = useRef<HTMLInputElement>();
     const [searchValue, setSearchValue] = useState("");
     const [searchHasValue, setSearchHasValue] = useState(false);
 
     async function openApp(url: string) {
-        setWeb(await encodeURL(url, service), webOpen, setWebOpen)
+        //@ts-ignore
+        setWeb(await chemicalEncode(url, service), webOpen, setWebOpen)
     }
 
     const handleChange = async (e: any) => {