+++ src/settings.ts
@@ -1,4 +1,3 @@
-const bare = new URL("/bare/", window.location.href);
 const obfuscate = true;
 const rammerhead = true;
 const github = "https://github.com/cognetwork-dev/Metallic";
@@ -6,4 +5,4 @@ const discord = "https://discord.com/invite/yk33HZSZkU";
 const privacyPolicyUpdated = 1707797140659;
 const gamesCdn = "https://radon.games";
 
-export { bare, obfuscate, rammerhead, github, discord, privacyPolicyUpdated, gamesCdn };
+export { obfuscate, rammerhead, github, discord, privacyPolicyUpdated, gamesCdn };